backend/requirements-local.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
Authlib==1.3.2
bcrypt==4.2.0
beautifulsoup4==4.12.3
ccxt==4.4.17
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
dateparser==1.2.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
fastapi-jwt-auth==0.5.0
fonttools==4.54.1
frozendict==2.4.5
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
kiwisolver==1.4.7
lxml==5.3.0
Mako==1.3.5
MarkupSafe==3.0.1
matplotlib==3.9.2
multidict==6.1.0
multitasking==0.0.11
numpy==2.0.2
packaging==24.1
pandas==2.2.3
passlib==1.7.4
peewee==3.17.6
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
propcache==0.2.0
pyasn1==0.6.1
pycares==4.4.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyJWT==1.7.1
pyparsing==3.1.4
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
pytz==2024.2
regex==2024.9.11
requests==2.32.3
rsa==4.9
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.38.6
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.31.1
webencodings==0.5.1
websockets==13.1
yarl==1.15.0
yfinance==0.2.44
zipp==3.20.2
